Allow passing of email as Session parameter
Generally, when creating a session, you either have to specify the
attribute on the form as being `:username`:

```ruby
= form.email_field :username
```

or adjust the params in the controller to explicitly assign the
`email` param to be `username`.

```ruby
def session_params
  {
    username: params[:session][:email],
    password: params[:session][:password],
  }
end
```

But when validating an account in the guts of Challah, `username`
and `email` are [used
interchangeably](https://github.com/jdtornow/challah/blob/master/lib/challah/concerns/user/findable.rb#L16).

This addition allows for someone to pass in `email` as a session param
as an alternative to `username`.